Moved Footer and header to components folder. Created base template for other pages to use. Created Event box component
