+++ communityApp/Screens/Components/Footer.js
